requirements.txt
openai
pyrogram
tgcrypto
pyyaml